[MISC][WK] Clean up the code
