+++ src/calendar/calendar-month.tsx
@@ -1,7 +1,8 @@
 import { useEffect, useState } from "react";
-import dayjs from "dayjs";
+import dayjs, { Dayjs } from "dayjs";
 import { MonthCell, MonthPickerContainer } from "./calendar-month.style";
 import { CalendarMonthProps, VariantMonth } from "./types";
+import { CalendarHelper } from "src/util/calendar-helper";
 
 export type Month = typeof Months[number];
 
@@ -11,7 +12,7 @@ export const CalendarMonth = ({
     showView,
     selectedStartDate,
 }: CalendarMonthProps) => {
-    const [monthDate, setMonthDate] = useState<string[]>([""]);
+    const [monthDate, setMonthDate] = useState<Dayjs[]>([]);
 
     useEffect(() => {
         if (showView === "Month") {
@@ -19,10 +20,10 @@ export const CalendarMonth = ({
         }
     }, [showView, calendarDate]);
 
-    const generateMonthStatus = (date: string) => {
-        const arrayDate = date.split("-");
+    const generateMonthStatus = (date: Dayjs) => {
+        const arrayDate = dayjs(date).format("YYYY-MMMM-DD").split("-");
         const month = arrayDate[1] as Month;
-        const value = dayjs(date).format("YYYY-MM-DD");
+        const value = date.format("YYYY-MM-DD");
         let variant: VariantMonth = "default";
 
         if (!dayjs(value).isValid())
@@ -46,32 +47,12 @@ export const CalendarMonth = ({
     };
 
     const generateMonth = () => {
-        const months: string[] = [];
-        const [yyyy, , dd] = calendarDate.format("YYYY-MM-DD").split("-");
-
-        for (let i = 0; i < 12; i++) {
-            const dayInMonth = dayjs(`${yyyy}-${Months[i]}-01`).daysInMonth();
-            let month = "";
-
-            month = `${yyyy}-${Months[i]}-${dd}`;
-
-            /**
-             * month would be 2023-01-30 but it not exist 30 in Feb
-             * dd === 30
-             * dayInMonth it may only 28 in the month
-             */
-            if (+dd > dayInMonth) {
-                // get last day of the month
-                month = `${yyyy}-${Months[i]}-${dayInMonth}`;
-            }
-
-            months.push(month);
-        }
+        const months = CalendarHelper.generateMonth(calendarDate);
 
         setMonthDate(months);
     };
 
-    if (monthDate.length < 1) return null;
+    if (monthDate.length === 1) return null;
 
     return (
         <MonthPickerContainer>
