+++ src/calendar/calendar-year.tsx
@@ -1,7 +1,8 @@
 import { useEffect, useState } from "react";
-import dayjs from "dayjs";
+import dayjs, { Dayjs } from "dayjs";
 import { CalendarYearProps, VariantYear } from "./types";
 import { YearCell, YearPickerContainer } from "./calendar-year.style";
+import { CalendarHelper } from "src/util/calendar-helper";
 
 export const CalendarYear = ({
     calendarDate,
@@ -9,7 +10,7 @@ export const CalendarYear = ({
     onClick,
     selectedStartDate,
 }: CalendarYearProps) => {
-    const [yearDate, setYearDate] = useState<string[]>(Array(12).fill(""));
+    const [yearDate, setYearDate] = useState<Dayjs[]>([]);
 
     useEffect(() => {
         if (showView === "Year") {
@@ -17,53 +18,32 @@ export const CalendarYear = ({
         }
     }, [showView, calendarDate]);
 
-    const generateYearStatus = (date: string) => {
+    const generateYearStatus = (date: Dayjs) => {
         const grayYearIndex = [0, 11];
-        const isDecadedYear = grayYearIndex.includes(yearDate.indexOf(date));
-        const year = date.split("-")[0];
+
+        const isDecadeYear = grayYearIndex.includes(yearDate.indexOf(date));
+        const fullDate = date.format("YYYY-MM-DD");
+        const year = fullDate.split("-")[0];
 
         let variant: VariantYear = "default";
 
-        variant = isDecadedYear
-            ? "nextDecaded"
-            : dayjs(selectedStartDate).isSame(date, "year")
+        variant = isDecadeYear
+            ? "nextDecade"
+            : dayjs(selectedStartDate).isSame(fullDate, "year")
             ? "selectedYear"
-            : dayjs().isSame(date, "year")
+            : dayjs().isSame(fullDate, "year")
             ? "currentYear"
             : "default";
 
         return {
             year,
-            value: date,
+            value: fullDate,
             variant: variant,
         };
     };
 
     const generateDecadeOfYear = () => {
-        const [yyyy, mm, dd] = calendarDate.format("YYYY-MM-DD").split("-");
-
-        const decaded = Math.floor(+yyyy / 10) * 10;
-        const years: string[] = [];
-
-        for (let i = 0; i <= 10; i++) {
-            let year = "";
-
-            if (i === 0) {
-                year = dayjs(`${decaded}-${mm}-${dd}`)
-                    .subtract(1, "year")
-                    .format("YYYY-MM-DD");
-
-                years.push(year, `${decaded}-${mm}-${dd}`);
-            }
-
-            if (i !== 0) {
-                year = dayjs(`${decaded}-${mm}-${dd}`)
-                    .add(i, "year")
-                    .format("YYYY-MM-DD");
-
-                years.push(year);
-            }
-        }
+        const years = CalendarHelper.generateDecadeOfYear(calendarDate);
 
         setYearDate(years);
     };
@@ -74,10 +54,6 @@ export const CalendarYear = ({
         <YearPickerContainer>
             {yearDate.map((date, yearIndex) => {
                 const { variant, year, value } = generateYearStatus(date);
-
-                if (variant === "selectedYear") {
-                    console.log("date: ", date);
-                }
 
                 return (
                     <YearCell
